@@ -8,13 +8,13 @@
     "lint": "vue-cli-service lint"
   },
   "dependencies": {
-    "bootstrap": "^4.6.1",
-    "bootstrap-vue": "^2.23.1",
+    "@mdi/font": "^7.3.67",
+    "axios": "^1.6.1",
     "core-js": "^3.8.3",
-    "popper.js": "^1.16.1-lts",
-    "portal-vue": "^2.1.7",
-    "vue": "^2.7.15",
-    "vue-simple-otp-input": "^1.9.0"
+    "vue": "^3.2.26",
+    "vue-router": "^4.2.5",
+    "vuetify": "^3.3.22",
+    "vuex": "^4.0.2"
   },
   "devDependencies": {
     "@babel/core": "^7.12.16",
@@ -23,8 +23,7 @@
     "@vue/cli-plugin-eslint": "~5.0.0",
     "@vue/cli-service": "~5.0.0",
     "eslint": "^7.32.0",
-    "eslint-plugin-vue": "^8.0.3",
-    "vue-template-compiler": "^2.6.14"
+    "eslint-plugin-vue": "^8.0.3"
   },
   "eslintConfig": {
     "root": true,
@@ -32,7 +31,7 @@
       "node": true
     },
     "extends": [
-      "plugin:vue/essential",
+      "plugin:vue/vue3-essential",
       "eslint:recommended"
     ],
     "parserOptions": {
@@ -43,6 +42,7 @@
   "browserslist": [
     "> 1%",
     "last 2 versions",
-    "not dead"
+    "not dead",
+    "not ie 11"
   ]
 }
