@@ -4,12 +4,14 @@ import org.springframework.context.annotation.Configuration;
 import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
 
-import com.nobody.OrderSmoothAPI.interceptor.helloInterceptor;
+import com.nobody.OrderSmoothAPI.interceptor.OwnerAccessInterceptor;
 
 @Configuration
 public class WebConfig implements WebMvcConfigurer {
+
     @Override
     public void addInterceptors(InterceptorRegistry registry) {
-        registry.addInterceptor(new helloInterceptor()).addPathPatterns("/hello1/**");
+        registry.addInterceptor(new OwnerAccessInterceptor()).addPathPatterns("/owner/**");
     }
+
 }
